+++ packages/material-ui-codemod/src/v5.0.0/variant-prop.test.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import { expect } from 'chai';
 import jscodeshift from 'jscodeshift';
-import transform from './textfield-variant-prop';
+import transform from './variant-prop';
 
 function read(fileName) {
   return fs.readFileSync(path.join(__dirname, fileName), 'utf8').toString();
@@ -10,18 +10,18 @@ function read(fileName) {
 
 describe('@material-ui/codemod', () => {
   describe('v5.0.0', () => {
-    describe('textfield-variant-prop', () => {
+    describe('variant-prop', () => {
       it('transforms props as needed', () => {
         const actual = transform(
           {
-            source: read('./textfield-variant-prop.test/actual.js'),
-            path: require.resolve('./textfield-variant-prop.test/actual.js'),
+            source: read('./variant-prop.test/actual.js'),
+            path: require.resolve('./variant-prop.test/actual.js'),
           },
-          { jscodeshift: jscodeshift },
+          { jscodeshift },
           {},
         );
 
-        const expected = read('./textfield-variant-prop.test/expected.js');
+        const expected = read('./variant-prop.test/expected.js');
         expect(actual).to.equal(expected, 'The transformed version should be correct');
       });
     });