+++ packages/material-ui/src/Switch/Switch.js
@@ -8,7 +8,7 @@ import { alpha, darken, lighten } from '../styles/colorManipulator';
 import capitalize from '../utils/capitalize';
 import SwitchBase from '../internal/SwitchBase';
 import useThemeProps from '../styles/useThemeProps';
-import experimentalStyled, { shouldForwardProp } from '../styles/experimentalStyled';
+import experimentalStyled from '../styles/experimentalStyled';
 import switchClasses, { getSwitchUtilityClass } from './switchClasses';
 
 const overridesResolver = (props, styles) => {
@@ -104,7 +104,7 @@ const SwitchRoot = experimentalStyled(
 
 const SwitchSwitchBase = experimentalStyled(
   SwitchBase,
-  { shouldForwardProp: (prop) => shouldForwardProp(prop) || prop === 'classes' },
+  {},
   {
     name: 'MuiSwitch',
     slot: 'SwitchBase',