+++ packages/material-ui/src/Typography/typography.spec.tsx
@@ -10,8 +10,8 @@ const TypographyTest = () => {
       <Typography align="left" color="initial" display="inline" />
       <Typography align="right" color="primary" display="initial" />
       <Typography align="justify" color="secondary" display="initial" />
-      <Typography align="inherit" color="textPrimary" />
-      <Typography align="inherit" color="textSecondary" />
+      <Typography align="inherit" color="text.primary" />
+      <Typography align="inherit" color="text.secondary" />
       <Typography align="inherit" color="error" />
       {/* TODO: system props did not catch this error. Add @ts-expect-error after it is fixed. */}
       <Typography display="incorrectValue" />